outreach-tool/requirements.txt
Flask==3.0.3
google-api-python-client==2.133.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
markdown==3.6
python-dotenv==1.0.1
gunicorn==22.0.0
PyYAML==6.0.1
pytest==8.0.0
pytest-timeout==2.2.0
requests==2.31.0